Move session initialization to the authentication store
