--- src/store/index.ts
@@ -1,4 +1,0 @@
-import { createPinia } from 'pinia'
-
-const store = createPinia()
-export default store
